BankApp.java

+++ src/main/java/BankApp.java
@@ -3,6 +3,8 @@ import java.text.NumberFormat;
 import java.util.Locale;
 import java.util.Scanner;
 
+import main.java.Bank;
+
 public class BankApp {
 
     public static void main(String[] args) {
@@ -12,7 +14,7 @@ public class BankApp {
             String ssn = scanner.nextLine();
             System.out.print("Enter password: ");
             String pw = scanner.nextLine();
-            CheckingAccount account = new CheckingAccount("6637862074333514921", 5000.00);
+            CheckingAccount account = new CheckingAccount("6637862074333514921", BigDecimal.valueOf(5000.00));
             BankApp.displayMenu(account, scanner);
 
         }
